Consolidate the funding pool onto the pool page
The pool page ({collective}/pool) now holds everything about the pool:
members enroll and withdraw there as before, and admins now open/reopen,
close, change the ceiling, and attach/detach funded agents there too —
including wind-down mode when pools lapse. The settings page keeps only
a pointer paragraph, and its controller stops loading pool state it no
longer renders.

All pool operations move out of CollectivesController into a new
FundingPoolsController, and every pool route now lives under the pool
prefix: the settings-prefixed pool endpoints and describe/execute
actions are deleted outright (nothing depends on them), with attach,
detach, and the new set_pool_ceiling action homed under pool/actions.
The pool actions index now evaluates conditional actions for the
viewer, so it lists exactly what the caller can do with the pool.

The ceiling leaves the settings write paths entirely: update_settings
and the update_collective_settings action no longer accept
member_daily_draw_cap — a dedicated update_ceiling endpoint serves the
pool page form and set_pool_ceiling serves agents, both gated on pool
availability. Pool tests move to funding_pools_controller_test.rb with
coverage for the new admin controls, no-pool page states, and the
responsibility split of the actions indexes.
