+++ app/views/collectives/settings.md.erb
@@ -87,24 +87,11 @@ Trio agents: <%= trio_personas.map { |a| "[#{a.name}](#{a.path}) ([task runs](/a
 <% end %>
 <% end %>
 
-<% if @funding_pool && !@funding_pool.archived? %>
+<% if @current_collective.standard? && @current_tenant.feature_enabled?("stripe_billing") && (@funding_pools_enabled || @funding_pool) %>
 
 ## Funding Pool
-<% unless @funding_pools_enabled %>
 
-**Funding pools are not available for this collective**, so no balances are being drawn on. This section is wind-down only: members can withdraw, agents can be detached, and the pool can be closed. Enrollment, attachment, and ceiling changes are paused until the collective is back on the paid plan or a platform admin enables the feature.
-<% end %>
-
-Each LLM call an attached agent makes is billed to one enrolled member's own prepaid balance, chosen at random per call. The pool never holds money. Enrolling is consenting to fund, with a ceiling the enrollee states themselves; enrollment and withdrawal are always self-serve on the [pool page](<%= @current_collective.path %>/pool), where every member can see the pool's state. Trio is on the pool's payroll automatically while the pool is open — it cannot be detached (disable Trio or close the pool instead). Attaching other agents (`attach_funded_agent` / `detach_funded_agent`) is admin-only, requires operator enablement, and the agent's principal must be enrolled.
-
-| Setting | Value |
-|---------|-------|
-| Pool draw ceiling | <%= format("$%.2f per member per %s", @funding_pool.member_draw_cap_cents / 100.0, @funding_pool.member_draw_cap_period) %> |
-| Enrolled members | <%= @pool_enrollments.any? ? @pool_enrollments.map { |e| "[#{e.user.display_name}](#{e.user.path})" }.join(", ") : "(none yet)" %> |
-<% pool_enrolled_user_ids = @pool_enrollments.map(&:user_id) -%>
-| Funded agents | <%= @funded_agents.any? ? @funded_agents.map { |a| "[#{a.display_name}](#{a.path})#{" (not running — principal not enrolled)" unless a.collective_pool_agent?(@funding_pool) || (a.parent_id && pool_enrolled_user_ids.include?(a.parent_id))}" }.join(", ") : "(none yet)" %> |
-<% my_enrollment = @current_user && @pool_enrollments.find { |e| e.user_id == @current_user.id } -%>
-| You | <%= my_enrollment ? format("Enrolled — the pool's agents can draw from your prepaid balance, up to your own ceiling of $%.2f per %s (or the pool's ceiling, whichever is lower)", my_enrollment.draw_cap_cents / 100.0, my_enrollment.draw_cap_period) : "Not enrolled" %> |
+A funding pool lets members fund this collective's agents together: each LLM call an attached agent makes is billed to one enrolled member's own prepaid balance. The pool — its state, enrollment, and admin management — lives on the [pool page](<%= @current_collective.path %>/pool)<%= @funding_pool ? "." : ", where an admin can open one." %>
 <% end %>
 
 <% if @current_tenant.feature_enabled?("stripe_billing") && !@current_collective.is_main_collective? && @current_collective.paid_tier? %>
