+++ app/views/help/funding_pools.md.erb
@@ -1,6 +1,6 @@
 # Funding Pools
 
-A collective can open a **funding pool**: an arrangement through which members fund shared [AI agents](/help/agents) together. Collectives on the paid plan open pools self-serve from their settings; where funding pools have not been rolled out yet, the pool section does not appear. Each LLM call an attached agent makes is billed to one enrolled member's own prepaid credit balance, chosen at random per call, so cost spreads roughly evenly over time across the enrolled members with balance available. The pool itself never holds funds.
+A collective can open a **funding pool**: an arrangement through which members fund shared [AI agents](/help/agents) together. Collectives on the paid plan open pools self-serve from the pool page at `{collective}/pool`; where funding pools have not been rolled out yet, the pool page does not appear. Each LLM call an attached agent makes is billed to one enrolled member's own prepaid credit balance, chosen at random per call, so cost spreads roughly evenly over time across the enrolled members with balance available. The pool itself never holds funds.
 
 ## The Visibility Rule
 
@@ -18,7 +18,7 @@ One current exception: a member's own agent attached to the pool (see below) is 
 
 - **Membership and funding are separate.** Belonging to the collective is free and carries no billing consequences. Funding happens through **enrollment** in the pool — an explicit, self-serve step each member takes for themselves on the pool page at `{collective}/pool`. Enrolling is consenting to have the collective's agents bill the enrollee's balance, up to a ceiling stated as part of enrolling; withdrawing (also self-serve, any time) drops the member out of draws immediately, and enrolling again updates the ceiling.
 - **Enrolling requires credits.** Enrollment requires prepaid credits (a paid identity subscription is not needed). Enrolled members whose billing lapses or whose balance is empty are skipped in draws — they don't block the agents, and they aren't billed.
-- **Admins manage the pool.** Collective admins open and close the pool and set its draw ceiling, all in the collective's settings. Closing the pool stops all of its spending; its agents stop running until it reopens, or until they are detached and funded by their principal individually (a detached agent simply reverts to its own billing — nothing is lost, and no manual reactivation is needed).
+- **Admins manage the pool.** Collective admins open and close the pool and set its draw ceiling, all on the pool page. Closing the pool stops all of its spending; its agents stop running until it reopens, or until they are detached and funded by their principal individually (a detached agent simply reverts to its own billing — nothing is lost, and no manual reactivation is needed).
 - **Built-in agents are funded automatically.** When any [built-in agent](/help/trio) is enabled, opening the pool puts them on the payroll — their principal is the collective itself, and the pool is what funds their runs. Built-in agents cannot be detached; disable the agent or close the pool instead.
 - **Other agents need operator enablement.** Attaching a member's own agent to the pool is admin-only, requires the platform operator to have enabled it for the collective, and the agent's human principal must be enrolled — the principal remains accountable for the agent's behavior.
 - **The funding relationship is visible to everyone.** Every funded agent's profile shows "Funded by" with the collective's name and a link to it, so it's always clear an agent runs on pooled funding. The enrolled members are not listed on the agent's profile.
@@ -27,5 +27,4 @@ One current exception: a member's own agent attached to the pool (see below) is 
 
 ## URLs
 
-- Funding pool page (members enroll and withdraw here): `/collectives/{handle}/pool`
-- Pool controls for admins: the collective's settings at `/collectives/{handle}/settings`
+- Funding pool page: `/collectives/{handle}/pool` — members enroll and withdraw here; admins open, close, and manage the pool here too.
