+++ config/routes.rb
@@ -581,27 +581,28 @@ Rails.application.routes.draw do
     post "#{prefix}/unarchive" => 'collectives#unarchive'
     post "#{prefix}/settings/add_ai_agent" => 'collectives#add_ai_agent'
     delete "#{prefix}/settings/remove_ai_agent" => 'collectives#remove_ai_agent'
-    get "#{prefix}/pool" => 'collectives#pool'
-    get "#{prefix}/pool/actions" => 'collectives#actions_index_pool'
-    get "#{prefix}/pool/actions/enroll_in_funding_pool" => 'collectives#describe_enroll_in_funding_pool'
-    post "#{prefix}/pool/actions/enroll_in_funding_pool" => 'collectives#execute_enroll_in_funding_pool'
-    get "#{prefix}/pool/actions/withdraw_from_funding_pool" => 'collectives#describe_withdraw_from_funding_pool'
-    post "#{prefix}/pool/actions/withdraw_from_funding_pool" => 'collectives#execute_withdraw_from_funding_pool'
-    post "#{prefix}/settings/create_funding_pool" => 'collectives#create_funding_pool'
-    post "#{prefix}/settings/close_funding_pool" => 'collectives#close_funding_pool'
-    post "#{prefix}/settings/enroll_in_funding_pool" => 'collectives#enroll_in_funding_pool'
-    delete "#{prefix}/settings/withdraw_from_funding_pool" => 'collectives#withdraw_from_funding_pool'
-    post "#{prefix}/settings/add_funded_agent" => 'collectives#add_funded_agent'
-    delete "#{prefix}/settings/remove_funded_agent" => 'collectives#remove_funded_agent'
+    # The pool page owns every pool operation — page, form endpoints, and
+    # described actions all live under the pool prefix.
+    get "#{prefix}/pool" => 'funding_pools#show'
+    get "#{prefix}/pool/actions" => 'funding_pools#actions_index'
+    get "#{prefix}/pool/actions/enroll_in_funding_pool" => 'funding_pools#describe_enroll_in_funding_pool'
+    post "#{prefix}/pool/actions/enroll_in_funding_pool" => 'funding_pools#execute_enroll_in_funding_pool'
+    get "#{prefix}/pool/actions/withdraw_from_funding_pool" => 'funding_pools#describe_withdraw_from_funding_pool'
+    post "#{prefix}/pool/actions/withdraw_from_funding_pool" => 'funding_pools#execute_withdraw_from_funding_pool'
+    get "#{prefix}/pool/actions/attach_funded_agent" => 'funding_pools#describe_attach_funded_agent'
+    post "#{prefix}/pool/actions/attach_funded_agent" => 'funding_pools#execute_attach_funded_agent'
+    get "#{prefix}/pool/actions/detach_funded_agent" => 'funding_pools#describe_detach_funded_agent'
+    post "#{prefix}/pool/actions/detach_funded_agent" => 'funding_pools#execute_detach_funded_agent'
+    get "#{prefix}/pool/actions/set_pool_ceiling" => 'funding_pools#describe_set_pool_ceiling'
+    post "#{prefix}/pool/actions/set_pool_ceiling" => 'funding_pools#execute_set_pool_ceiling'
+    post "#{prefix}/pool/create_funding_pool" => 'funding_pools#create_funding_pool'
+    post "#{prefix}/pool/close_funding_pool" => 'funding_pools#close_funding_pool'
+    post "#{prefix}/pool/enroll_in_funding_pool" => 'funding_pools#enroll_in_funding_pool'
+    delete "#{prefix}/pool/withdraw_from_funding_pool" => 'funding_pools#withdraw_from_funding_pool'
+    post "#{prefix}/pool/add_funded_agent" => 'funding_pools#add_funded_agent'
+    delete "#{prefix}/pool/remove_funded_agent" => 'funding_pools#remove_funded_agent'
+    post "#{prefix}/pool/update_ceiling" => 'funding_pools#update_ceiling'
     get "#{prefix}/settings/actions" => 'collectives#actions_index_settings'
-    get "#{prefix}/settings/actions/enroll_in_funding_pool" => 'collectives#describe_enroll_in_funding_pool'
-    post "#{prefix}/settings/actions/enroll_in_funding_pool" => 'collectives#execute_enroll_in_funding_pool'
-    get "#{prefix}/settings/actions/withdraw_from_funding_pool" => 'collectives#describe_withdraw_from_funding_pool'
-    post "#{prefix}/settings/actions/withdraw_from_funding_pool" => 'collectives#execute_withdraw_from_funding_pool'
-    get "#{prefix}/settings/actions/attach_funded_agent" => 'collectives#describe_attach_funded_agent'
-    post "#{prefix}/settings/actions/attach_funded_agent" => 'collectives#execute_attach_funded_agent'
-    get "#{prefix}/settings/actions/detach_funded_agent" => 'collectives#describe_detach_funded_agent'
-    post "#{prefix}/settings/actions/detach_funded_agent" => 'collectives#execute_detach_funded_agent'
     get "#{prefix}/settings/actions/update_collective_settings" => 'collectives#describe_update_collective_settings'
     post "#{prefix}/settings/actions/update_collective_settings" => 'collectives#update_collective_settings_action'
     get "#{prefix}/settings/actions/add_ai_agent_to_collective" => 'collectives#describe_add_ai_agent_to_collective'
